Fix tagging issue
If a channel which not include a title tag in the stream is played after a channel
which correctly updates the title, the old title tag will persist.

Try adding the Russian Radio Maximum (stream: http://maximum.fmtuner.ru) and play
that stream AFTER a stream known to include the correct title tag. When switching
to Radio Maximum, the old title will show after a while.

The issues is fixed by resetting the tag variable.